// src/modules/recommendation/types/questions.enum.ts
export enum RecommendationQuestionEnum {
  ALLERGY = 'Are you allergic to any of the following food items?',
  DIET = 'Do you identify as any of the following?',
  FOOD_TYPE = 'What types of food does your family enjoy eating?',
  FLAVORS = 'Does your family typically prefer dishes with the following flavors:',
  COOKING_METHOD = 'How does your family prefer dishes to be prepared?',
  MEAL_TIME = 'How long does a typical meal of your family usually last?',
  BEVERAGE = 'What type of beverage do our family prefer?',
  FASTFOOD = 'How often does your family eat fast food?',
  CUISINE = 'Which culinary traditions do you prefer?',
  SPICE_LEVEL = 'Does your family enjoy eating spicy food?',
}